Apps/UCG-Max-Webhook-Receiver/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
apscheduler==3.10.4
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1